requirements.txt
streamlit==1.32.0
pandas==2.2.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.1